Split hotspot self-heal probes and policy
