+++ apps/server/vibesensor/adapters/hotspot/self_heal.py
@@ -8,29 +8,19 @@ from __future__ import annotations
 
 import logging
 import subprocess
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Protocol
 
-from vibesensor.adapters.hotspot.parsers import (
-    HealStateStore,
-    expected_ip_match,
-    nm_log_signals,
-    parse_active_conn_device,
-    parse_active_connection_names,
-    parse_ip,
-    parse_iw_info,
-    parse_port53_conflict,
-    parse_rfkill_blocked,
+from vibesensor.adapters.hotspot.health_probe import (
+    HealthState,
+    collect_health,
+    journalctl_nm_args,
 )
+from vibesensor.adapters.hotspot.parsers import HealStateStore
+from vibesensor.adapters.hotspot.remediation import apply_heals
 
 LOGGER = logging.getLogger("vibesensor.adapters.hotspot.selfheal")
-
-_DHCP_BAD_SIGNALS: frozenset[str] = frozenset(
-    {"dhcp_no_range", "dhcp_dnsmasq_start_failed", "port53_conflict"},
-)
-
-_FALLBACK_CHANNELS: tuple[int, ...] = (1, 6, 11)
 
 
 class HotspotSelfHealConfig(Protocol):
@@ -81,283 +71,6 @@ class CommandRunner:
             return CommandResult(returncode=124, stdout="", stderr=f"timeout: {' '.join(argv)}")
 
 
-@dataclass(slots=True)
-class HealAction:
-    """Record of a single self-heal action taken to restore hotspot connectivity."""
-
-    name: str
-    detected: str
-    action: str
-    helped: bool
-
-
-@dataclass(slots=True)
-class HealthState:
-    """Snapshot of the current Wi-Fi hotspot health collected from system commands."""
-
-    nm_running: bool
-    wifi_radio_on: bool
-    rfkill_blocked: bool
-    iface_exists: bool
-    iface_up: bool
-    ap_conn_exists: bool
-    ap_conn_active: bool
-    ap_mode: bool
-    ip_ok: bool
-    dhcp_ok: bool
-    channel: str | None
-    last_error_category: str
-    dhcp_log_signal: str | None = None
-    port53_conflict: str | None = None
-    issues: list[str] = field(default_factory=list)
-
-    @property
-    def ok(self) -> bool:
-        return (
-            self.nm_running
-            and self.wifi_radio_on
-            and not self.rfkill_blocked
-            and self.iface_exists
-            and self.iface_up
-            and self.ap_conn_exists
-            and self.ap_conn_active
-            and self.ap_mode
-            and self.ip_ok
-            and self.dhcp_ok
-        )
-
-
-def _journalctl_nm_args(lookback_minutes: int) -> list[str]:
-    """Build journalctl argv for NetworkManager log retrieval."""
-    return [
-        "journalctl",
-        "-u",
-        "NetworkManager",
-        "--since",
-        f"-{max(1, lookback_minutes)} min",
-        "--no-pager",
-        "-n",
-        "120",
-    ]
-
-
-def _find_port53_conflict(runner: CommandRunner) -> str | None:
-    ss = runner.run(["ss", "-ltnup", "sport", "=", ":53"], timeout_s=5)
-    if ss.returncode != 0:
-        return None
-    return parse_port53_conflict(ss.stdout)
-
-
-def collect_health(
-    ap: HotspotApConfig,
-    self_heal: HotspotSelfHealConfig,
-    runner: CommandRunner,
-) -> HealthState:
-    """Collect the current hotspot health state by running diagnostic commands."""
-    issues: list[str] = []
-
-    nm_active = runner.run(["systemctl", "is-active", "NetworkManager"], timeout_s=5)
-    nm_running = nm_active.returncode == 0 and nm_active.stdout.strip() == "active"
-    if not nm_running:
-        issues.append("networkmanager_down")
-
-    wifi_state = runner.run(["nmcli", "-t", "-f", "WIFI", "general", "status"], timeout_s=5)
-    wifi_radio_on = wifi_state.returncode == 0 and wifi_state.stdout.strip().lower() == "enabled"
-    if not wifi_radio_on:
-        issues.append("wifi_radio_off")
-
-    rfkill_blocked = False
-    rfkill_check = runner.run(["rfkill", "list"], timeout_s=5)
-    if rfkill_check.returncode == 0:
-        rfkill_blocked = parse_rfkill_blocked(rfkill_check.stdout)
-    if rfkill_blocked:
-        issues.append("rfkill_blocked")
-
-    iface = runner.run(["ip", "link", "show", "dev", ap.ifname], timeout_s=5)
-    iface_exists = iface.returncode == 0
-    iface_up = iface_exists and " state UP " in f" {iface.stdout} "
-    if not iface_exists:
-        issues.append("iface_missing")
-    elif not iface_up:
-        issues.append("iface_down")
-
-    con_list = runner.run(["nmcli", "-t", "-f", "NAME", "connection", "show"], timeout_s=5)
-    con_names = parse_active_connection_names(con_list.stdout)
-    ap_conn_exists = ap.con_name in con_names
-    if not ap_conn_exists:
-        issues.append("ap_connection_missing")
-
-    active = runner.run(
-        ["nmcli", "-t", "-f", "NAME,DEVICE", "connection", "show", "--active"],
-        timeout_s=5,
-    )
-    ap_conn_active, active_device = parse_active_conn_device(ap.con_name, active.stdout)
-    if not ap_conn_active:
-        issues.append("ap_connection_inactive")
-    elif active_device and active_device != ap.ifname:
-        issues.append("ap_active_on_wrong_if")
-
-    iw_info = runner.run(["iw", "dev", ap.ifname, "info"], timeout_s=5)
-    ap_mode = False
-    channel = None
-    if iw_info.returncode == 0:
-        ap_mode, channel = parse_iw_info(iw_info.stdout)
-    if ap_conn_active and not ap_mode:
-        issues.append("iface_not_ap_mode")
-
-    ip_show = runner.run(["ip", "-4", "addr", "show", "dev", ap.ifname], timeout_s=5)
-    actual_ip = parse_ip(ip_show.stdout) if ip_show.returncode == 0 else None
-    ip_ok = expected_ip_match(ap.ip, actual_ip)
-    if not ip_ok:
-        issues.append("ip_mismatch")
-
-    nm_logs = runner.run(
-        _journalctl_nm_args(self_heal.diagnostics_lookback_minutes),
-        timeout_s=8,
-    )
-    dhcp_log_signal, _ = nm_log_signals(nm_logs.stdout)
-
-    pgrep_dnsmasq = runner.run(["pgrep", "-af", "dnsmasq"], timeout_s=5)
-    nm_dnsmasq_running = pgrep_dnsmasq.returncode == 0 and any(
-        "networkmanager" in line.lower() and "dnsmasq" in line.lower()
-        for line in pgrep_dnsmasq.stdout.splitlines()
-    )
-    dhcp_ok = nm_dnsmasq_running and dhcp_log_signal not in _DHCP_BAD_SIGNALS
-    if not dhcp_ok:
-        issues.append("dhcp_unhealthy")
-
-    conflict = _find_port53_conflict(runner)
-    if conflict:
-        issues.append("port53_conflict")
-
-    last_error = issues[-1] if issues else "none"
-
-    return HealthState(
-        nm_running=nm_running,
-        wifi_radio_on=wifi_radio_on,
-        rfkill_blocked=rfkill_blocked,
-        iface_exists=iface_exists,
-        iface_up=iface_up,
-        ap_conn_exists=ap_conn_exists,
-        ap_conn_active=ap_conn_active,
-        ap_mode=ap_mode,
-        ip_ok=ip_ok,
-        dhcp_ok=dhcp_ok,
-        channel=channel,
-        last_error_category=last_error,
-        dhcp_log_signal=dhcp_log_signal,
-        port53_conflict=conflict,
-        issues=issues,
-    )
-
-
-def _ensure_ap_connection(
-    ap: HotspotApConfig, runner: CommandRunner, channel: int | None = None
-) -> bool:
-    configured_channel = channel if channel is not None else ap.channel
-    con_names = parse_active_connection_names(
-        runner.run(["nmcli", "-t", "-f", "NAME", "connection", "show"], timeout_s=8).stdout,
-    )
-    if not ap.psk and ap.con_name in con_names:
-        # Recreate open AP profiles to guarantee no stale security fields remain.
-        runner.run(["nmcli", "connection", "delete", ap.con_name], timeout_s=8)
-        con_names = [name for name in con_names if name != ap.con_name]
-
-    if ap.con_name not in con_names:
-        added = runner.run(
-            [
-                "nmcli",
-                "connection",
-                "add",
-                "type",
-                "wifi",
-                "ifname",
-                ap.ifname,
-                "con-name",
-                ap.con_name,
-                "autoconnect",
-                "yes",
-                "ssid",
-                ap.ssid,
-            ],
-            timeout_s=10,
-        )
-        if added.returncode != 0:
-            return False
-
-    modify_args = [
-        "nmcli",
-        "connection",
-        "modify",
-        ap.con_name,
-        "802-11-wireless.mode",
-        "ap",
-        "802-11-wireless.band",
-        "bg",
-        "802-11-wireless.channel",
-        str(configured_channel),
-        "ipv4.method",
-        "shared",
-        "ipv4.addresses",
-        ap.ip,
-        "ipv6.method",
-        "ignore",
-    ]
-    if ap.psk:
-        modify_args.extend(
-            [
-                "802-11-wireless-security.key-mgmt",
-                "wpa-psk",
-                "802-11-wireless-security.psk",
-                ap.psk,
-            ]
-        )
-
-    modified = runner.run(modify_args, timeout_s=10)
-    if modified.returncode != 0:
-        return False
-
-    up = runner.run(["nmcli", "--wait", "12", "connection", "up", ap.con_name], timeout_s=15)
-    return up.returncode == 0
-
-
-def _bounce_connection(ap: HotspotApConfig, runner: CommandRunner) -> None:
-    runner.run(["nmcli", "connection", "down", ap.con_name], timeout_s=8)
-    runner.run(["ip", "link", "set", ap.ifname, "up"], timeout_s=5)
-    runner.run(["nmcli", "--wait", "10", "connection", "up", ap.con_name], timeout_s=12)
-
-
-def _recreate_connection(ap: HotspotApConfig, runner: CommandRunner) -> bool:
-    runner.run(["nmcli", "connection", "delete", ap.con_name], timeout_s=8)
-    return _ensure_ap_connection(ap, runner)
-
-
-def _handle_port53_conflict(
-    conflict: str,
-    runner: CommandRunner,
-) -> str:
-    lowered = conflict.lower()
-    if "dnsmasq" in lowered:
-        runner.run(["systemctl", "disable", "--now", "dnsmasq.service"], timeout_s=10)
-        return "stopped standalone dnsmasq service"
-
-    if "systemd-resolve" in lowered:
-        runner.run(
-            [
-                "/bin/sh",
-                "-c",
-                "mkdir -p /etc/systemd/resolved.conf.d && "
-                "printf '[Resolve]\\nDNSStubListener=no\\n' > "
-                "/etc/systemd/resolved.conf.d/vibesensor-no-stub.conf",
-            ],
-            timeout_s=10,
-        )
-        runner.run(["systemctl", "restart", "systemd-resolved"], timeout_s=10)
-        return "disabled systemd-resolved DNS stub listener"
-
-    return f"detected :53 conflict owner={conflict}; no automatic disruptive action taken"
-
-
 def _emit_diagnostics(
     ap: HotspotApConfig,
     lookback_minutes: int,
@@ -372,7 +85,7 @@ def _emit_diagnostics(
         ["ip", "addr", "show", "dev", ap.ifname],
         ["iw", "dev", ap.ifname, "info"],
         ["rfkill", "list"],
-        _journalctl_nm_args(lookback_minutes),
+        journalctl_nm_args(lookback_minutes),
     ]
 
     logger.warning("hotspot diagnostics begin")
@@ -421,132 +134,7 @@ def run_self_heal_once(
     _log_summary("degraded", ap, health)
     _emit_diagnostics(ap, self_heal.diagnostics_lookback_minutes, runner, LOGGER)
 
-    actions: list[HealAction] = []
-    if not health.nm_running:
-        if state_store.allow("restart_networkmanager", self_heal.min_restart_interval_seconds):
-            runner.run(["systemctl", "restart", "NetworkManager"], timeout_s=15)
-            actions.append(
-                HealAction(
-                    name="restart_networkmanager",
-                    detected="NetworkManager inactive",
-                    action="systemctl restart NetworkManager",
-                    helped=False,
-                ),
-            )
-        else:
-            actions.append(
-                HealAction(
-                    name="restart_networkmanager",
-                    detected="NetworkManager inactive",
-                    action="restart skipped by backoff",
-                    helped=False,
-                ),
-            )
-
-    if not health.wifi_radio_on:
-        runner.run(["nmcli", "radio", "wifi", "on"], timeout_s=8)
-        actions.append(
-            HealAction(
-                name="wifi_radio_on",
-                detected="Wi-Fi radio disabled",
-                action="nmcli radio wifi on",
-                helped=False,
-            ),
-        )
-
-    if health.rfkill_blocked:
-        runner.run(["rfkill", "unblock", "wifi"], timeout_s=8)
-        actions.append(
-            HealAction(
-                name="rfkill_unblock",
-                detected="rfkill blocked",
-                action="rfkill unblock wifi",
-                helped=False,
-            ),
-        )
-
-    if not health.iface_up and health.iface_exists:
-        runner.run(["ip", "link", "set", ap.ifname, "up"], timeout_s=8)
-        actions.append(
-            HealAction(
-                name="if_up",
-                detected="interface down",
-                action=f"ip link set {ap.ifname} up",
-                helped=False,
-            ),
-        )
-
-    if not health.ap_conn_exists or not health.ap_conn_active:
-        ensured = _ensure_ap_connection(ap, runner)
-        if not ensured:
-            ensured = _recreate_connection(ap, runner)
-        if not ensured:
-            for fallback_channel in _FALLBACK_CHANNELS:
-                if fallback_channel == ap.channel:
-                    continue
-                if _ensure_ap_connection(ap, runner, channel=fallback_channel):
-                    actions.append(
-                        HealAction(
-                            name="ap_channel_fallback",
-                            detected="configured AP channel failed",
-                            action=f"ap recreated on fallback channel {fallback_channel}",
-                            helped=False,
-                        ),
-                    )
-                    ensured = True
-                    break
-        actions.append(
-            HealAction(
-                name="ensure_ap_connection",
-                detected="AP connection missing/inactive",
-                action="ensure AP connection and bring it up",
-                helped=False,
-            ),
-        )
-
-    if health.ap_conn_active and (not health.iface_up or not health.ap_mode):
-        _bounce_connection(ap, runner)
-        actions.append(
-            HealAction(
-                name="bounce_ap",
-                detected="AP active but interface down or not in AP mode",
-                action="nmcli connection down/up and ip link up",
-                helped=False,
-            ),
-        )
-
-    if not health.dhcp_ok:
-        if health.port53_conflict:
-            message = _handle_port53_conflict(health.port53_conflict, runner)
-            actions.append(
-                HealAction(
-                    name="port53_conflict",
-                    detected=f"port 53 conflict ({health.port53_conflict})",
-                    action=message,
-                    helped=False,
-                ),
-            )
-        _ensure_ap_connection(ap, runner)
-        if state_store.allow("restart_networkmanager", self_heal.min_restart_interval_seconds):
-            runner.run(["systemctl", "restart", "NetworkManager"], timeout_s=15)
-            runner.run(["nmcli", "--wait", "12", "connection", "up", ap.con_name], timeout_s=15)
-            actions.append(
-                HealAction(
-                    name="dhcp_repair",
-                    detected="DHCP path unhealthy",
-                    action="re-applied AP connection and restarted NetworkManager",
-                    helped=False,
-                ),
-            )
-        else:
-            actions.append(
-                HealAction(
-                    name="dhcp_repair",
-                    detected="DHCP path unhealthy",
-                    action="restart skipped by backoff; AP re-applied",
-                    helped=False,
-                ),
-            )
+    actions = apply_heals(ap, self_heal, health, runner, state_store)
 
     healed = collect_health(ap, self_heal, runner)
     status = "healed" if healed.ok else "failed"
